MultiTrajectory: Switch KF to MultiTrajectory
The gain matrix smoother and updator now accept MJ and TSProxies respectively.
The KF is changed to accomodate this. The KF now does not pre-fill measurements
into track states, but appends them to the MJ as it goes along. Calibration is
now done by the KF itself, the GainMatrixUpdator loses the calibrator object and
doesn't to it anymore.

WARNING: The Updator currently uses dynamic matrices! This is slow and will be
replaced with fixed size matrix arithmetic again!

The Updator and Smoother tests are updated to support the new arguments, but
expect the same calculation results as before.